chore: extract sn_node crate
