+++ sn_node/Cargo.toml
@@ -1,9 +1,13 @@
 [package]
-name = "safenode"
-version = "0.1.0"
+authors = ["MaidSafe Developers <dev@maidsafe.net>"]
+description = "Safe Node"
+name = "sn_node"
+version = "0.83.0"
 edition = "2021"
 license = "GPL-3.0"
 homepage = "https://maidsafe.net"
+readme = "README.md"
+repository = "https://github.com/maidsafe/safe_network"
 
 [[bin]]
 name = "safenode"
