+++ sn_node/examples/registers.rs
