+++ sn_node/src/bin/safenode/rpc.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use safenode::node::RunningNode;
+use sn_node::RunningNode;
 
 use super::NodeCtrl;
