+++ sn_node/src/error.rs
