+++ sn_node/src/lib.rs
@@ -43,8 +43,31 @@
 #[macro_use]
 extern crate tracing;
 
-#[cfg(test)]
-mod tests_e2e;
+mod api;
+mod error;
+mod event;
 
-/// SAFE Node
-pub mod node;
+pub use self::{
+    api::RunningNode,
+    event::{NodeEvent, NodeEventsChannel, NodeEventsReceiver},
+};
+
+use self::api::TransferAction;
+
+use libp2p::{Multiaddr, PeerId};
+use sn_domain::{node_transfers::Transfers, storage::RegisterStorage};
+use sn_networking::Network;
+use tokio::sync::mpsc;
+
+/// `Node` represents a single node in the distributed network. It handles
+/// network events, processes incoming requests, interacts with the data
+/// storage, and broadcasts node-related events.
+pub struct Node {
+    network: Network,
+    registers: RegisterStorage,
+    transfers: Transfers,
+    events_channel: NodeEventsChannel,
+    /// Peers that are dialed at startup of node.
+    initial_peers: Vec<(PeerId, Multiaddr)>,
+    transfer_actor: mpsc::Sender<TransferAction>,
+}
